# src/main/java/com/piere/bootcamp/bankaccounts/model/dto/OperationDto.java
package com.piere.bootcamp.bankaccounts.model.dto;

import java.io.Serializable;

import lombok.Data;

@Data
public class OperationDto implements Serializable {
    
    private static final long serialVersionUID = 1L;

    private Double amount;
}